+++ src/machinable/execution/execution.py
@@ -7,6 +7,7 @@ from typing import Any, Callable, Union
 import pendulum
 import yaml
 
+from expandvars import expandvars
 from ..config.interface import ConfigInterface, mapped_config
 from ..core.exceptions import ExecutionException
 from ..core.settings import get_settings
@@ -134,32 +135,6 @@ class Execution(Jsonable):
 
     def set_experiment(self, experiment):
         self.experiment = Experiment.create(experiment)
-
-        # experiment directory
-        experiment_directory = self.experiment.specification.get("directory", None)
-        if experiment_directory is None:
-            experiment_directory = get_settings()["default_directory"]
-        if isinstance(experiment_directory, str):
-            # replace magic variables
-            if experiment_directory.find("&PROJECT") != -1:
-                project_name = self.project.name if self.project else ""
-                experiment_directory = experiment_directory.replace(
-                    "&PROJECT", project_name.replace(".", "/")
-                )
-            if experiment_directory.find("&MODULE") != -1:
-                module_name = (
-                    self.experiment._resolved_module_origin
-                    if hasattr(self.experiment, "_resolved_module_origin")
-                    else ""
-                )
-                experiment_directory = experiment_directory.replace(
-                    "&MODULE", module_name
-                )
-            try:
-                experiment_directory = dt.now().strftime(experiment_directory)
-            except ValueError:
-                pass
-            self.storage.config["directory"] = experiment_directory.strip("/")
 
         return self
 
@@ -315,6 +290,48 @@ class Execution(Jsonable):
 
         self.failures = 0
         self.storage.config["experiment"] = self.experiment_id
+
+        # expand variables in storage directory
+        if isinstance(self.storage.config["directory"], str):
+            # simplify if proposed expandvars changes are accepted
+            # https://github.com/sayanarijit/expandvars/issues/25
+            # https://github.com/sayanarijit/expandvars/issues/26
+
+            _environ_project = os.environ.get("PROJECT")
+            try:
+                project_name = self.project.name if self.project else None
+            except (KeyError, ValueError):
+                project_name = None
+            if project_name is not None:
+                os.environ["PROJECT"] = project_name
+            _environ_experiment = os.environ.get("EXPERIMENT")
+            experiment_name = (
+                self.experiment.specification["name"] if self.experiment else ""
+            )
+            if experiment_name is not None:
+                os.environ["EXPERIMENT"] = experiment_name
+
+            self.storage.config["directory"].replace("&", "$")
+            self.storage.config["directory"] = expandvars(
+                self.storage.config["directory"]
+            )
+
+            if _environ_project is not None:
+                os.environ["PROJECT"] = _environ_project
+            if _environ_experiment is not None:
+                os.environ["EXPERIMENT"] = _environ_experiment
+            # ---
+
+            # strftime variables
+            try:
+                self.storage.config["directory"] = dt.now().strftime(
+                    self.storage.config["directory"]
+                )
+            except ValueError:
+                pass
+            self.storage.config["directory"] = self.storage.config["directory"].strip(
+                "/"
+            )
 
         # check if URL is an existing experiment
         derived_from = None